requirements.txt
pynacl >= 1.4.0
